tests are now correctly located in the packages/cli directory structure, which is consistent with the rest of the API implementation
