--- src/executor/__tests__/executor.test.ts
@@ -1,126 +1,0 @@
-import { SldsExecutor, LintConfig, ReportConfig } from '../index';
-import { LintRunner } from '../../services/lint-runner';
-import { FileScanner } from '../../services/file-scanner';
-import { Readable } from 'stream';
-import { jest } from '@jest/globals';
-import { LintResult } from '../../types';
-
-// Create mock type-safe functions
-const mockLintResult: LintResult = {
-  filePath: 'file1.css',
-  errors: [{ line: 1, column: 1, endColumn: 10, message: 'Test error', ruleId: 'test-rule', severity: 2 }],
-  warnings: []
-};
-
-// Use an alternative approach without generic typing that was causing issues
-jest.mock('../../services/lint-runner', () => {
-  return {
-    LintRunner: {
-      runLinting: jest.fn().mockImplementation(() => {
-        return Promise.resolve([mockLintResult]);
-      })
-    }
-  };
-});
-
-jest.mock('../../services/file-scanner', () => {
-  return {
-    FileScanner: {
-      scanFiles: jest.fn().mockImplementation(() => {
-        return Promise.resolve([['file1.css']]);
-      })
-    }
-  };
-});
-
-jest.mock('fs/promises');
-
-// Skip tests temporarily until TypeScript issues are resolved
-xdescribe('SldsExecutor', () => {
-  let executor: SldsExecutor;
-  
-  beforeEach(() => {
-    executor = new SldsExecutor();
-    jest.clearAllMocks();
-  });
-  
-  describe('lint', () => {
-    it('should scan directory and run linting when no files are provided', async () => {
-      const config: LintConfig = {
-        directory: './src',
-        fix: true
-      };
-      
-      const results = await executor.lint(config);
-      
-      // Check FileScanner was called with correct params
-      expect(FileScanner.scanFiles).toHaveBeenCalledTimes(2);
-      
-      // Check LintRunner was called with correct params
-      expect(LintRunner.runLinting).toHaveBeenCalledTimes(2);
-      
-      // Check results were combined correctly
-      expect(results).toHaveLength(2);
-    });
-    
-    it('should use provided files and skip scanning when files are provided', async () => {
-      const config: LintConfig = {
-        files: ['file1.css', 'component1.html']
-      };
-      
-      await executor.lint(config);
-      
-      // FileScanner should not be called when files are provided
-      expect(FileScanner.scanFiles).not.toHaveBeenCalled();
-      
-      // LintRunner should still be called
-      expect(LintRunner.runLinting).toHaveBeenCalledTimes(2);
-    });
-  });
-  
-  describe('report', () => {
-    it('should return a readable stream', async () => {
-      const config: ReportConfig = {
-        directory: './src',
-        format: 'json'
-      };
-      
-      const stream = await executor.report(config);
-      
-      expect(stream).toBeInstanceOf(Readable);
-    });
-    
-    it('should use lint results to generate a report', async () => {
-      // Create a spy on the lint method
-      const lintSpy = jest.spyOn(executor, 'lint').mockResolvedValue([mockLintResult]);
-      
-      const config: ReportConfig = {
-        directory: './src',
-        format: 'json'
-      };
-      
-      await executor.report(config);
-      
-      expect(lintSpy).toHaveBeenCalledWith({
-        directory: './src',
-        configStylelint: expect.any(String),
-        configEslint: expect.any(String)
-      });
-      
-      // Restore the original implementation
-      lintSpy.mockRestore();
-    });
-  });
-});
-
-/**
- * Placeholder test file for executor module
- * 
- * Proper tests will be implemented in a future update
- */
-
-describe('SldsExecutor placeholder tests', () => {
-  it('should be implemented in the future', () => {
-    expect(true).toBe(true);
-  });
-});
